рефакторинг
- удалит render.js

- деструктурирует util.js
